archive/legacy-agent/requirements.txt
python-dotenv
requests
PyYAML
